ref #6 Function for removing tree, Record DTO type
Record DTO type is represented by type Record and RecordList.
Record is an alternative for Tree, but without the pointers to children.
RecordList is a wrapper structure that holds an array of Records and
its size.

This structures and the related functions will be used for transfering
data from the StringGrid object to an array, sorting the array and
generating a new tree that is based on this data.